auth part2

+++ flask_api/resources/item.py
@@ -1,7 +1,7 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_api.db import db
 from flask_api.models import ItemModel
 from flask_api.schemas import ItemSchema, ItemUpdateSchema
@@ -18,6 +18,9 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -45,7 +48,8 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
-    @jwt_required()  # when using jwt_required the API request need to contain a header "Authorization":"Bearer Token"
+    # when using jwt_required the API request need to contain a header "Authorization":"Bearer Token"
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
